requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
markdown==3.7
gunicorn==23.0.0
